Add comprehensive template debugging to identify data flow issue
- Enhanced prompt.php debugging to verify $curriculum_template value at multiple points
- Added validation checks in template extraction to confirm data has content
- Improved ai_service.php logging to show template data summary at each step
- Added logging to generate_module_with_template wrapper function to verify it's called
- Improved build_template_prompt_guidance() logging to show full guidance text
- Fixed courseid type conversion in template_reader get_curriculum_templates()
- Validates template structure/activities before passing to AI

+++ classes/local/template_reader.php
@@ -54,8 +54,8 @@ class template_reader {
             $parts = explode('|', $line);
             if (count($parts) >= 2) {
                 $name = trim($parts[0]);
-                $courseid = trim($parts[1]);
-                $sectionid = isset($parts[2]) ? trim($parts[2]) : null;
+                $courseid = (int)trim($parts[1]);
+                $sectionid = isset($parts[2]) ? (int)trim($parts[2]) : null;
                 
                 // Validate course exists and user has access
                 if ($this->validate_template_access($courseid)) {
